Organizes existing models/views/serializers into folder structure
