+++ nftgram/models/follows_model.py
@@ -1,0 +1,11 @@
+from django.db import models
+from .user_model import User
+
+
+class Follows(models.Model):
+
+    user_id = models.ManyToManyField(User, related_name="+")
+    follows_id = models.ManyToManyField(User, related_name="+")
+
+    class Meta:
+        db_table = "follows"
