+++ nftgram/serializers/post_serializers.py
@@ -1,0 +1,17 @@
+from rest_framework import serializers
+from nftgram.models import Post
+
+
+class PostSerializer(serializers.HyperlinkedModelSerializer):
+
+    class Meta:
+        model = Post
+        fields = (
+            'post_id',
+            'user_id',
+            'title',
+            'description',
+            'posted_at',
+            'nft_id',
+            'nft_url',
+        )
