+++ nftgram/urls.py
@@ -1,11 +1,17 @@
 from django.urls import path, re_path, include
-from nftgram.views import UserViewSet
+from nftgram.views import UserViewSet, PostViewSet, ProfileViewSet, FollowsViewSet, ReplyViewSet
 from rest_framework import routers
 from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
+
 
 # Routers provide an easy way of automatically determining the URL conf.
 router = routers.DefaultRouter()
 router.register(r'users', UserViewSet)
+router.register(r'posts', PostViewSet)
+router.register(r'profiles', ProfileViewSet)
+router.register(r'follows', FollowsViewSet)
+router.register(r'replies', ReplyViewSet)
+
 
 urlpatterns = [
     path('', include(router.urls)),
