+++ nftgram/views/__init__.py
@@ -1,0 +1,7 @@
+from .user_views import UserViewSet
+from .post_views import PostViewSet
+from .profile_views import ProfileViewSet
+from .follows_views import FollowsViewSet
+from .reply_views import ReplyViewSet
+
+
